// src/models/ClientEmployeesModel.tsx
export interface ClientEmployeesModel {
    Id: string;
    ClientId: string;
    ClientBusinessUnitId: string; 
    Name:string;
    Mobile: number;
    Email: string; 
    Department :string;
    Location : string;
    Designation : string;
   // ID; ClientID; ClientBusinessUnitID; Name;Mobile; email; Department (text); Location(text); Designation(text)
}